**First change**: Decode only the new tokens as they are generated during streaming instead of the entire sequence. This makes a difference when the context is large (like 40,000 tokens).

Here is a test with ExLlama_HF and a 13b model. The prompt is cached to make the difference more pronounced.

Before:

```
Output generated in 10.28 seconds (48.56 tokens/s, 499 tokens, context 3550, seed 1197861167)
Output generated in 10.27 seconds (48.59 tokens/s, 499 tokens, context 3550, seed 1463481238)
Output generated in 10.30 seconds (48.43 tokens/s, 499 tokens, context 3550, seed 1587775594)
```

After:

```
Output generated in 9.69 seconds (51.50 tokens/s, 499 tokens, context 3550, seed 909082919)
Output generated in 9.71 seconds (51.41 tokens/s, 499 tokens, context 3550, seed 221584078)
Output generated in 9.72 seconds (51.36 tokens/s, 499 tokens, context 3550, seed 105310110)
```

**Second change**: Restrict the UI streaming updates to 5 per second when `--listen` or `--share` are provided. I find that this makes the UI a lot more responsive over a network.